[Nexters/covid-letter-web#29] ♻️ 구글,네이버 로그인 컴포넌트 분리
